tests: implement tool argument casting
This is just for manual tests to see if this is the only problem.
LLMs should provide correct types for tool calls and the agent
(in our case the tests implement the agent) should not need to
cast the arguments.
